Comments
Above is just general feedback on the practices and conventions to adopt which I have learned (like global constant should be in caps eg. VARIABLE_NAME). I never check your algorithm so cannot comment if objectives are met. =) I'm not familiar with Mocha testing framework so not able to feedback on your test file.